feat: ReservationPolicy 추가
