+++ src/main/java/roomescape/domain/ReservationDate.java
@@ -5,13 +5,9 @@ import roomescape.exception.BusinessException;
 import roomescape.exception.ErrorCode;
 
 import java.time.LocalDate;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Getter
 public class ReservationDate {
-
-    private static final int RESERVABLE_DAYS_RANGE = 14;
 
     private final LocalDate date;
 
@@ -32,10 +28,5 @@ public class ReservationDate {
 
     public boolean isToday() {
         return date.isEqual(LocalDate.now());
-    }
-
-    public static List<LocalDate> getReservableDates() {
-        final LocalDate today = LocalDate.now();
-        return today.datesUntil(today.plusDays(RESERVABLE_DAYS_RANGE)).collect(Collectors.toList());
     }
 }
